+++ Engine/DataFeeds/SubscriptionSynchronizer.cs
@@ -17,9 +17,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using NodaTime;
 using QuantConnect.Data.UniverseSelection;
-using QuantConnect.Securities;
 
 namespace QuantConnect.Lean.Engine.DataFeeds
 {
@@ -28,10 +26,9 @@ namespace QuantConnect.Lean.Engine.DataFeeds
     /// </summary>
     public class SubscriptionSynchronizer : ISubscriptionSynchronizer, ITimeProvider
     {
-        private ITimeProvider _timeProvider;
-        private readonly CashBook _cashBook;
-        private DateTimeZone _sliceTimeZone;
         private readonly UniverseSelection _universeSelection;
+        private readonly TimeSliceFactory _timeSliceFactory;
+        private ITimeProvider _timeProvider;
         private ManualTimeProvider _frontierTimeProvider;
 
         /// <summary>
@@ -44,13 +41,13 @@ namespace QuantConnect.Lean.Engine.DataFeeds
         /// </summary>
         /// <param name="universeSelection">The universe selection instance used to handle universe
         /// selection subscription output</param>
-        /// <param name="cashBook">The cash book, used for creating the cash book updates</param>
+        /// <param name="timeSliceFactory">Used to create the new <see cref="TimeSlice"/></param>
         /// <returns>A time slice for the specified frontier time</returns>
         public SubscriptionSynchronizer(UniverseSelection universeSelection,
-                                        CashBook cashBook)
+                                        TimeSliceFactory timeSliceFactory)
         {
             _universeSelection = universeSelection;
-            _cashBook = cashBook;
+            _timeSliceFactory = timeSliceFactory;
         }
 
         /// <summary>
@@ -65,19 +62,6 @@ namespace QuantConnect.Lean.Engine.DataFeeds
             }
             _timeProvider = timeProvider;
             _frontierTimeProvider = new ManualTimeProvider(_timeProvider.GetUtcNow());
-        }
-
-        /// <summary>
-        /// Sets the time zone of the created slice object. If already set will throw.
-        /// </summary>
-        /// <param name="timeZone">The time zone of the created slice object</param>
-        public void SetSliceTimeZone(DateTimeZone timeZone)
-        {
-            if (_sliceTimeZone != null)
-            {
-                throw new Exception("SubscriptionSynchronizer.SetSliceTimeZone(): can only be called once");
-            }
-            _sliceTimeZone = timeZone;
         }
 
         /// <summary>
@@ -197,7 +181,7 @@ namespace QuantConnect.Lean.Engine.DataFeeds
             }
             while (newChanges != SecurityChanges.None);
 
-            var timeSlice = TimeSlice.Create(frontierUtc, _sliceTimeZone, _cashBook, data, changes, universeDataForTimeSliceCreate);
+            var timeSlice = _timeSliceFactory.Create(frontierUtc, data, changes, universeDataForTimeSliceCreate);
 
             return timeSlice;
         }
