+++ Tests/Engine/DataFeeds/FileSystemDataFeedTests.cs
@@ -54,7 +54,7 @@ namespace QuantConnect.Tests.Engine.DataFeeds
                 marketHoursDatabase);
             algorithm.SubscriptionManager.SetDataManager(dataManager);
             var synchronizer = new Synchronizer();
-            synchronizer.Initialize(algorithm, dataManager, false, algorithm.Portfolio.CashBook);
+            synchronizer.Initialize(algorithm, dataManager, false);
 
             feed.Initialize(algorithm, job, resultHandler, mapFileProvider, factorFileProvider, dataProvider, dataManager, synchronizer);
             algorithm.Initialize();
