+++ Tests/Engine/DataFeeds/SubscriptionSynchronizerTests.cs
@@ -77,9 +77,9 @@ namespace QuantConnect.Tests.Engine.DataFeeds
             var endTimeUtc = algorithm.EndDate.ConvertToUtc(TimeZones.NewYork);
             var startTimeUtc = algorithm.StartDate.ConvertToUtc(TimeZones.NewYork);
             var subscriptionBasedTimeProvider = new SubscriptionFrontierTimeProvider(startTimeUtc, dataManager);
-            var synchronizer = new SubscriptionSynchronizer(dataManager.UniverseSelection, algorithm.Portfolio.CashBook);
+            var timeSliceFactory = new TimeSliceFactory{TimeZone = algorithm.TimeZone};
+            var synchronizer = new SubscriptionSynchronizer(dataManager.UniverseSelection, timeSliceFactory);
             synchronizer.SetTimeProvider(subscriptionBasedTimeProvider);
-            synchronizer.SetSliceTimeZone(algorithm.TimeZone);
             var totalDataPoints = 0;
             var subscriptions = dataManager.DataFeedSubscriptions;
             foreach (var kvp in algorithm.Securities)
