+++ Tests/Engine/RealTimePriceUpdateTests.cs
@@ -17,17 +17,18 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using NodaTime;
 using NUnit.Framework;
 using QuantConnect.Algorithm;
 using QuantConnect.Brokerages;
 using QuantConnect.Data;
 using QuantConnect.Data.Auxiliary;
+using QuantConnect.Data.Market;
 using QuantConnect.Data.UniverseSelection;
 using QuantConnect.Lean.Engine.DataFeeds;
 using QuantConnect.Lean.Engine.Results;
 using QuantConnect.Packets;
 using QuantConnect.Securities;
+using QuantConnect.Tests.Common.Securities;
 
 namespace QuantConnect.Tests.Engine
 {
@@ -36,6 +37,7 @@ namespace QuantConnect.Tests.Engine
     {
         private TestableLiveTradingDataFeed _liveTradingDataFeed;
         private SecurityExchangeHours _exchangeHours;
+        private SubscriptionDataConfig _config;
 
         [TestFixtureSetUp]
         public void Setup()
@@ -73,10 +75,12 @@ namespace QuantConnect.Tests.Engine
                 marketHoursDatabase);
             algo.SubscriptionManager.SetDataManager(dataManager);
             var synchronizer = new Synchronizer();
-            synchronizer.Initialize(algo, dataManager, true, algo.Portfolio.CashBook);
+            synchronizer.Initialize(algo, dataManager, true);
             _liveTradingDataFeed.Initialize(algo, jobPacket, new LiveTradingResultHandler(), new LocalDiskMapFileProvider(),
                                             null, new DefaultDataProvider(), dataManager, synchronizer);
             algo.Initialize();
+
+            _config = SecurityTests.CreateTradeBarConfig();
         }
 
         /// <summary>
@@ -98,7 +102,7 @@ namespace QuantConnect.Tests.Engine
         [Test]
         public void NullSubscriptions_DoNotIndicateRealTimePriceUpdates()
         {
-            Assert.IsFalse(_liveTradingDataFeed.UpdateRealTimePrice(null, new TimeZoneOffsetProviderNeverOpen()));
+            Assert.IsFalse(_liveTradingDataFeed.UpdateRealTimePrice(null, new TimeZoneOffsetProviderNeverOpen(), _exchangeHours));
         }
 
         [Test]
@@ -111,19 +115,10 @@ namespace QuantConnect.Tests.Engine
                 SymbolProperties.GetDefault("USA"),
                 ErrorCurrencyConverter.Instance
             );
-            var config = new SubscriptionDataConfig(
-                security.GetType(),
-                security.Symbol,
-                Resolution.Daily,
-                DateTimeZone.Utc,
-                _exchangeHours.TimeZone,
-                false,
-                false,
-                false
-            );
-            var subscriptionRequest = new SubscriptionRequest(false, null, security, config, DateTime.MinValue, DateTime.MaxValue);
+
+            var subscriptionRequest = new SubscriptionRequest(false, null, security, _config, DateTime.MinValue, DateTime.MaxValue);
             var subscription = new Subscription(subscriptionRequest, null, new TimeZoneOffsetProviderNeverOpen());
-            Assert.IsFalse(_liveTradingDataFeed.UpdateRealTimePrice(subscription, new TimeZoneOffsetProviderNeverOpen()));
+            Assert.IsFalse(_liveTradingDataFeed.UpdateRealTimePrice(subscription, new TimeZoneOffsetProviderNeverOpen(), _exchangeHours));
         }
 
         [Test]
@@ -136,26 +131,20 @@ namespace QuantConnect.Tests.Engine
                 SymbolProperties.GetDefault("USA"),
                 ErrorCurrencyConverter.Instance
             );
-            var config = new SubscriptionDataConfig(
-                security.GetType(),
-                security.Symbol,
-                Resolution.Daily,
-                DateTimeZone.Utc,
-                _exchangeHours.TimeZone,
-                false,
-                false,
-                false
-            );
-            var subscriptionRequest = new SubscriptionRequest(false, null, security, config, DateTime.MinValue, DateTime.MaxValue);
+
+            var subscriptionRequest = new SubscriptionRequest(false, null, security, _config, DateTime.MinValue, DateTime.MaxValue);
             var subscription = new Subscription(subscriptionRequest, null, new TimeZoneOffsetProviderNeverOpen());
-            Assert.IsTrue(_liveTradingDataFeed.UpdateRealTimePrice(subscription, new TimeZoneOffsetProviderAlwaysOpen()));
+            Assert.IsTrue(_liveTradingDataFeed.UpdateRealTimePrice(subscription, new TimeZoneOffsetProviderAlwaysOpen(), _exchangeHours));
         }
 
         class TestableLiveTradingDataFeed : LiveTradingDataFeed
         {
-            public bool UpdateRealTimePrice(Subscription subscription, TimeZoneOffsetProvider timeZoneOffsetProvider)
+            public bool UpdateRealTimePrice(
+                Subscription subscription,
+                TimeZoneOffsetProvider timeZoneOffsetProvider,
+                SecurityExchangeHours exchangeHours)
             {
-                return SubscriptionShouldUpdateRealTimePrice(subscription, timeZoneOffsetProvider);
+                return UpdateSubscriptionRealTimePrice(subscription, timeZoneOffsetProvider, exchangeHours, new Tick());
             }
         }
