+++ Tests/Python/PandasConverterTests.cs
@@ -480,11 +480,9 @@ namespace QuantConnect.Tests.Engine.DataFeeds
         {
             var subscriptionDataConfig = GetSubscriptionDataConfig<T>(symbol, resolution);
             var security = GetSecurity(subscriptionDataConfig);
-
-            return data.Select(t => TimeSlice.Create(
+            var timeSliceFactory = new TimeSliceFactory{ TimeZone = TimeZones.Utc };
+            return data.Select(t => timeSliceFactory.Create(
                t.Time,
-               TimeZones.Utc,
-               new CashBook(),
                new List<DataFeedPacket> { new DataFeedPacket(security, subscriptionDataConfig, new List<BaseData>() { t as BaseData }) },
                new SecurityChanges(Enumerable.Empty<Security>(), Enumerable.Empty<Security>()),
                 new Dictionary<Universe, BaseDataCollection>()).Slice);
